+++ client/src/components/Comment/Modal.styled.ts
@@ -76,3 +76,29 @@ export const Star=styled.div`
     }
   }
 `
+
+export const StarArr=styled.div`
+display:flex;
+gap:5px;`
+export const OneStar=styled.div`
+display:flex;
+border:1px solid white;
+flex-direction:row-reverse
+`
+export const LeftStar=styled.svg`
+  width: 15px;
+  height: 30px;
+  &:hover{
+    >path{
+      fill: aliceblue;
+    }
+    & ~svg{
+      >path{
+      fill: aliceblue;
+    }
+    }
+    
+  }
+`
+export const RightStar=styled(LeftStar)`
+`
